requirements.txt
pyyaml>=6.0
click>=8.0
rich>=13.0
python-dotenv>=1.0
requests>=2.28